requirements.txt
openai==1.9.0
python-dotenv==1.0.1
pymongo==4.6.1
certifi==2023.11.17
ttkbootstrap==1.10.1